Test harness for EPS files.
We now have a test harness for EPS files. Two variants were created one for the default scale=1 and one where scale=2. These two tests are run against two different EPS files, one with zero for the start of the bounding box and one where this is not the case. PNG test renders are used to make sure the output match what we expect. Lastly the sample EPS files were generated using the included create_eps.gnuplot file and the gnuplot program on a Mac.