Prefer homebrew freetype over X11 freetype (but still allow both)
I've recently included Pillow with a py2app build of a frozen application on OS X. When Pillow is installed on a machine that has X11, the preference for X11's libfreetype causes a new dependency for my frozen app.  I don't want my users to be required to install X11 if they don't have to (it's not included by default after OS X 10.8).

This PR adds a preference for homebrew's libfreetype (if available), which py2app detects and includes, and which doesn't create an X11 dependency in apps that are frozen and use Pillow (PIL).