No-seek implementation of HttpResponse Image.open
This is an alternative implementation for `Image.open()`
to support opening of images from HttpResponse objects. It
uses `.seekable()` to determine whether the file supports
seeking (with failover for this attribute not being supported
e.g. in `urllib2` response objects).

The prefix is chopped off as normal and pre-pended back onto
the bytes fed into the `io.IOBytes` object. This avoids loading
the image object totally (duplicating it in memory) until
we are sure we have a factory for the file.

Add AttributeError exception catch; add io.IOBytes wrapper only once

Add callable .seek test for Python2.x