test_imagecms: Reduce precision of extended info
In 32bit computers (when making wheels) it was found that the
precision of the extended information in test_extended_information() was
found to be lesser than what was being tested. This commit adds a helper
function which reduces the precision by truncating (multiplying by
10 ** digits, truncating to int and dividing by 10 ** digits) and then
testing the rtruncated numbers rather than the original numbers.