Add outline feature to ImageDraw.text()
This change adds an additional method argument to the draw() method
which draws the text in black, before drawing it in the color selected
in the fill parameter.

Additionally due to the positional requirements enforced on the method
by using **kwargs, move to using an explicit signature with default values.

* Helps improve readability
* Allows for easier modification later.