Trim trailing whitespace from all source files throughout project
Many editors trim whitespace by default. When newcomers are submitting
patches, it is often undesirable to see inadvertent whitespace changes
in the diff as it is distracting and unrelated. To make contributing
easier, clean up all whitespace in one go across the project. Will allow
for cleaner diffs in the future. Contributors will not need to worry
about automatic yet inadvertent whitespace changes.