Remove EOL (or nearly EOL) distro scripts from depends/
The distro scripts in the depends directory are unmaintained. Som
reference long EOL distributions. For example, Fedora 23 went EOL
2016-12-20. These scripts aren't referenced in the documentation either.

With the planned removal of Python2 support, most of these scripts will
be outdated and incompatible as well.

This removes the last use of install_openjpeg.sh, so that has been
removed as well.