src: PIL: JpegImagePlugin.py: make -1 and keep subsamplings identical for save
Make the -1 subsampling behave like "keep" for the save method.

The doc claims that -1 is equivalent to keep. It is actually equivalent
to 4:2:0 for some reason.

Note that -1 was representing both the -1 subsampling and the absence of
it. To persist the current behavior which is compatible with multiple
input file formats other than JPEG (which is not the case of the
subsampling parameter), make the absence of subsampling parameter equal
to -2.

get_sampling() remains unchanged since it is a public API.
Note that subsampling can be -1 and -2 in some cases, both results in an
image with a subsampling of 4:2:0.

Also removing faulty test case since subsampling -1 cannot be used on a
generated image but only JPEG file format.

Signed-off-by: Quentin Schulz <foss@0leil.net>