APNG - use previous frame with dispose action to draw the next frame.
APNG defines a dispose operation after each frame. This dispose should
be done before drawing the next frame. PNG code stores this frame on _prev_im
but it is not used to draw the next frame.
This change use the _prev_im as the base to draw the next frame.