Remove redundant wheel dep from pyproject.toml
Remove the redundant `wheel` dependency, as it is added by the backend
automatically.  Listing it explicitly in the documentation was
a historical mistake and has been fixed since, see:
https://github.com/pypa/setuptools/commit/f7d30a9529378cf69054b5176249e5457aaf640a

While Pillow uses a custom backend that modifies the `bdist_wheel`
method, it does not import `wheel` or use it in a way that would rely
on setuptools implementation details.