fix(avif): use only aomenc and dav1d to reduce wheel size
Also build libavif as a shared library. When it is built as a static
library, the dependency library files are combined into a single
archive. But when they are linked as a shared library, the linker is
able to remove unused objects. This yields a modest but not
insignificant file size reduction.