Fix assertion observed when adding new colors to an indexed-color TIFF
When ImagePalette._new_color_index adds a new color to an indexed-color
TIFF, an assertion added in f2302ab716e9e could be tripped. The call
into Image.histogram could result in ImageFile.load being called, which
could eventually replace ImagePalette._palette with a bytes object,
which did not satisfy the assertion that it be only bytearray.